+++ Changelog.md
@@ -1,3 +1,11 @@
+### dev
+[full changelog](http://github.com/rspec/rspec-core/compare/v2.12.2...master)
+
+Bug fixes
+
+* Don't blow up when dumping error output for instances
+  of anonymous error classes (Myron Marston).
+
 ### 2.12.2 / 2012-12-13
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.12.1...v2.12.2)