+++ Changelog.md
@@ -20,6 +20,8 @@ Enhancements
 * Add support for percent deltas to `be_within` matcher:
   `expect(value).to be_within(10).percent_of(expected)`
   (Myron Marston).
+* Add support to `include` matcher to allow it to be given a list
+  of matchers as the expecteds to match against (Luke Redpath).
 
 ### 2.12.1 / 2012-12-15
 [full changelog](http://github.com/rspec/rspec-expectations/compare/v2.12.0...v2.12.1)